# Faster Dictionary clone