Resolves https://github.com/dotnet/runtime/issues/41939

Fast `.Clone()` extension method can be made on top of this change as
```csharp
public static class DictionaryExtensions
{
    public static Dictionary<TKey, TValue> Clone<TKey, TValue>(this Dictionary<TKey, TValue> source) where TKey : notnull
    {
        return new Dictionary<TKey, TValue>(source, source.Comparer);
    }
}
```

String keys are the biggest improvement since they cost most to rehash

|                   Method | NetFx | InitialCount |      Mean | Allocated | Ratio | Change |
|------------------------- |------ |-------------:|----------:|----------:|------:|-------:|
|           Clone_Int_Full | NetFx |         3000 |  62.09 us |   66.1 KB |  2.45 |  x 0.4 |
|           Clone_Int_Full |  Base |         3000 |  25.32 us |   65.9 KB |  1.00 |      - |
|           Clone_Int_Full |    **PR** |         3000 |  10.56 us |   65.9 KB |  **0.42** |  **x 2.4** |
|                          |       |              |           |           |       |        |
|    Clone_Int_HalfRemoved | NetFx |         3000 |  32.43 us |   31.4 KB |  2.73 |  x 0.4 |
|    Clone_Int_HalfRemoved |  Base |         3000 |  11.88 us |   31.3 KB |  1.00 |      - |
|    Clone_Int_HalfRemoved |    **PR** |         3000 |   6.86 us |   31.3 KB |  **0.58** |  **x 1.7** |
|                          |       |              |           |           |       |        |
|        Clone_String_Full | NetFx |         3000 | 111.84 us |   92.4 KB |  1.86 |  x 0.5 |
|        Clone_String_Full |  Base |         3000 |  60.01 us |   92.3 KB |  1.00 |      - |
|        Clone_String_Full |    **PR** |         3000 |  26.84 us |   92.3 KB |  **0.45** |  **x 2.2** |
|                          |       |              |           |           |       |        |
| Clone_String_HalfRemoved | NetFx |         3000 |  58.25 us |   43.9 KB |  1.39 |  x 0.7 |
| Clone_String_HalfRemoved |  Base |         3000 |  42.04 us |   43.8 KB |  1.00 |      - |
| Clone_String_HalfRemoved |    **PR** |         3000 |  14.73 us |   43.8 KB |  **0.35** |  **x 2.9** |

https://github.com/dotnet/performance/compare/master...benaadams:DictionaryClone?expand=1