@benaadams 2020-09-08
Broken compress variant, but I'm quite impressed there are tests for these scenarios :)