@jkotas 2020-09-08
It would be nice to share performance numbers for the different cases.