@danmosemsft 2020-09-08
> @danmosemsft it would be nice if we also made this improvement in DictionarySlim. For specific cache-like use cases we roll our own dictionary and it would be nice to not have to do that.

@mjsabby the copy constructor specifically? It would be easy to add if someone would offer a PR. 

I am unsure what to do with DictionarySlim. It's encouraging that you're using it, or plan to. That may suggest we should remove the "Experimental" or even put it into the core collections. But @GrabYourPitchforks had concerns about it's "unsafeness" (I think related to mutating the entry in place). Also the name may suggest to people it's optimized for small number of items, which it's not (I happened to just run across @benaadams [type](https://github.com/benaadams/Ben.DictionarySlim) with the same name)