@benaadams 2020-09-11
>  all while continuing not to have evidence that this is actually a hot path for real-world apps. 

Signals Rosyln would move away from ImmuntableDictionary for a fast Dictionary clone method https://github.com/dotnet/roslyn/pull/47503#issuecomment-687917289

> Yes, ImmutableDictionary is really bad. We just haven't had cases yet that fell into the patterns for which we have production-ready alternatives. The main scenario that we have the ability to improve today is collections with the following characteristics:
>
>* The dictionary is built up from empty, as opposed to built by transforming a previous collection through add/remove operations
>* The collection has more than a few elements, but less than ~10K elements
>
>For this we could switch to an immutable wrapper around `Dictionary<TKey, TValue>`, similar to how we use `ImmutableArray<T>` instead of `ImmutableList<T>`.

/cc @sharwell

and

> My gut tells me this should be a slam dunk win. The vast majority (probably nearly all) of cases where we use ImmutableDict, it's not to benefit from what ID is decent at, but just to ensure it will not be mutated post creation. For that purpose it's actually a pretty terrible choice and adds a tremendous amount of overhead for functionality that we're not needing in those scenarios.

/cc @CyrusNajmabadi

Also for lockless read Dictionary, with very infrequent updates (where reallocating the entire dictionary is more lightweight than taking a lock on every read, or using `ConcurrentDictionary`) e.g.
``` C#
class LookupTable
{
    private Dictionary<int, string>? _lookUp;
    private readonly object _lock = new ();

    public bool TryGetValue(int key, [MaybeNullWhen(false)] out string value)
    {
        // No locks required for read
        var lookUp = _lookUp;
        if (lookUp is null)
        {
            value = default;
            return false;
        }

        return lookUp.TryGetValue(key, out value);
    }

    public void Add(int key, string value)
    {
        lock (_lock)
        {
            // Clone + Add
            var dict = _lookUp?.Clone() ?? new (capacity: 1);
            dict.Add(key, value);
            // Switch reference
            _lookUp = dict;
        }
    }

    public void AddRange(IEnumerable<KeyValuePair<int, string>> collection)
    {
        lock (_lock)
        {
            // Clone + Add
            var dict = _lookUp?.Clone() ?? new(capacity: 8);
            foreach (var entry in collection)
            {
                dict.Add(entry.Key, entry.Value);
            }
            // Switch reference
            _lookUp = dict;
        }
    }
}
```