@benaadams 2020-09-11
Its also why I wanted to add a clone extension method in  #41939
```csharp
public static class DictionaryExtensions
{
    public static Dictionary<TKey, TValue> Clone<TKey, TValue>(this Dictionary<TKey, TValue> source) where TKey : notnull
    {
        return new Dictionary<TKey, TValue>(source, source.Comparer);
    }
}
```
Because it's not clear that passing a `Dictionary` as an `IDictionary` to the `Dictionary` constructor is a fast way to copy it and on Full Framework it really isn't https://referencesource.microsoft.com/#mscorlib/system/collections/generic/dictionary.cs,106