@jkotas 2020-09-11
I believe that Roslyn is on .NET Standard 2.0 and needs to have great perf on .NET Framework for foreseeable future. If this is important for Roslyn perf, I would expect Roslyn to switch to a local copy of slim Dictionary that is both leaner than Dictionary and can be optimized  on .NET Framework too.