@benaadams 2020-09-11
> I believe that Roslyn is on .NET Standard 2.0 and needs to have great perf on .NET Framework for foreseeable future.

On .NET Framework this method does presize the dictionary before adding so it's likely to be a win in terms of allocations; whereas the allocations from their use of ImmutableDictionary is eye watering

![image](https://user-images.githubusercontent.com/1142958/92336163-93e0d480-f095-11ea-9f0d-2a297e09b04b.png)