@CyrusNajmabadi 2020-09-11
So, FWIW, i was saying that roslyn likely almost never needs to use ImmutableDict, and could instead just use a non-mutating view over a normal Dictionary.  I imagine the majority cases are that we don't need clone capabilities, as we likely don't actually do any sort of copy-on-write here.  I may be wrong though.  @benaadams were there particular hostpots where we could benefit from a readonly-dictionary, but also need to clone?